Ajout de JsFunctionProvider
appel de jsFunctionProvider dans Main
